requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
attrs
bcrypt
black
boto3
botocore
cachetools
certifi
cffi
charset-normalizer
click
cryptography
distro
dnspython
ecdsa
email-validator
# emergentintegrations==0.1.0 --extra-index-url https://d33sy5i8bnduwe.cloudfront.net/simple/
fastapi
fastuuid
filelock
flake8
frozenlist
fsspec
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
hf-xet
httpcore
httplib2
httptools
httpx
huggingface_hub
idna
importlib_metadata
iniconfig
isort
Jinja2
jiter
jmespath
jq
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
langchain
langchain-core
langchain-openai
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
litellm
markdown-it-py
MarkupSafe
mccabe
mdurl
motor
multidict
mypy
mypy_extensions
numpy
oauthlib
openai
orjson
ormsgpack
packaging
pandas
passlib
pathspec
pillow
platformdirs
pluggy
propcache
proto-plus
protobuf
pyasn1
pyasn1_modules
pycodestyle
pycparser
pydantic[email]
pydantic-settings
pydantic_core
pyflakes
Pygments
PyJWT
pymongo
pyparsing
pytest
pytest-asyncio
python-dateutil
python-dotenv
python-jose
python-multipart
pytokens
pytz
PyYAML
referencing
regex
requests
requests-oauthlib
requests-toolbelt
rich
rpds-py
rsa
s3transfer
s5cmd
shellingham
six
sniffio
SQLAlchemy
starlette
stripe
tiktoken
tokenizers
tqdm
typer
typer-slim
typing-inspection
typing_extensions
tzdata
uritemplate
urllib3
uvicorn
uvloop
watchfiles
xxhash
yarl
zipp
zstandard
fastmcp>=2.13.0.1
langgraph-cli[inmem]
langchain_mcp_adapters
langchain_anthropic
langchain_community
deepagents
nest_asyncio
autoflake
isort
black
aiofiles
composio
composio-langchain
langchain
langchain-openai
vapi_python
openai-agents
composio-openai-agents
